+++ common/core/src/main/java/zingg/common/core/block/Block.java
@@ -3,10 +3,8 @@ package zingg.common.core.block;
 import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
